Fix test directory cleanup

+++ tests/conftest.py
@@ -105,21 +105,24 @@ def spark_session():
 def make_test_project_dir_module():
     tmpdir = _make_project_dir(TEST_PROJECT_REPO)
     yield tmpdir / "dsgrid_project"
-    shutil.rmtree(tmpdir)
+    if tmpdir.exists():
+        shutil.rmtree(tmpdir)
 
 
 @pytest.fixture
 def make_test_project_dir():
     tmpdir = _make_project_dir(TEST_PROJECT_REPO)
     yield tmpdir / "dsgrid_project"
-    shutil.rmtree(tmpdir)
+    if tmpdir.exists():
+        shutil.rmtree(tmpdir)
 
 
 @pytest.fixture
 def make_standard_scenarios_project_dir():
     tmpdir = _make_project_dir(TEST_STANDARD_SCENARIOS_PROJECT_REPO)
     yield tmpdir / "dsgrid_project"
-    shutil.rmtree(tmpdir)
+    if tmpdir.exists():
+        shutil.rmtree(tmpdir)
 
 
 @pytest.fixture(scope="module")
@@ -131,19 +134,15 @@ def make_test_data_dir_module():
     dst_path = tmpdir / "datasets"
     shutil.copytree(Path(TEST_DATASET_DIRECTORY), dst_path)
     yield dst_path
-    shutil.rmtree(tmpdir)
+    if tmpdir.exists():
+        shutil.rmtree(tmpdir)
 
 
 @pytest.fixture
-def make_test_data_dir():
-    tmpdir = Path(gettempdir()) / "test_data"
-    if os.path.exists(tmpdir):
-        shutil.rmtree(tmpdir)
-    os.mkdir(tmpdir)
-    dst_path = tmpdir / "datasets"
+def make_test_data_dir(tmp_path):
+    dst_path = tmp_path / "datasets"
     shutil.copytree(Path(TEST_DATASET_DIRECTORY), dst_path)
     yield dst_path
-    shutil.rmtree(tmpdir)
 
 
 def _make_project_dir(project):
